simplify DSP request handlers and create a request handler for the /campaigns/:id route
