+++ client/lib/promote-post/index.ts
@@ -14,6 +14,12 @@ import {
 	isJetpackMinimumVersion,
 } from 'calypso/state/sites/selectors';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
+const DSP_ERROR_NO_LOCAL_USER = 'no_local_user';
+const DSP_URL_CHECK_UPSERT_USER = '/user/check';
+
+type NewDSPUserResult = {
+	new_dsp_user: boolean;
+};
 
 declare global {
 	interface Window {
@@ -203,6 +209,32 @@ export const requestDSP = async < T >(
 			return await wpcom.req.del( params );
 		default:
 			return await wpcom.req.get( params );
+	}
+};
+
+const handleDSPError = async < T >(
+	error: any,
+	siteId: number,
+	callerURl: string
+): Promise< T > => {
+	if ( error.errorCode === DSP_ERROR_NO_LOCAL_USER ) {
+		const createUserQuery = await requestDSP< NewDSPUserResult >(
+			siteId,
+			DSP_URL_CHECK_UPSERT_USER
+		);
+		if ( createUserQuery.new_dsp_user ) {
+			// then we should retry the original query
+			return await requestDSP< T >( siteId, callerURl );
+		}
+	}
+	throw error;
+};
+
+export const requestDSPHandleErrors = async < T >( siteId: number, url: string ): Promise< T > => {
+	try {
+		return await requestDSP( siteId, url );
+	} catch ( e ) {
+		return await handleDSPError( e, siteId, url );
 	}
 };
